+++ benchmark/cmd/benchmark/function.go
@@ -5,119 +5,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"sync"
 	"strconv"
+	"sync"
 	"time"
 )
-
-// -----------------------------------------------
-type ImageScaleRequest struct {
-	Bucket      string `json:"bucket"`
-	Source      string `json:"source"`
-	Destination string `json:"destination"`
-	ForceRemote bool   `json:"force_remote"`
-	ForceBackup bool   `json:"force_backup"`
-}
-
-type ImageScaleResponse struct {
-	ForceRemote      bool    `json:"force_remote"`
-	CodeDuration     float64 `json:"code_duration"`
-	DownloadDuration float64 `json:"download_duration"`
-	ScaleDuration    float64 `json:"scale_duration"`
-	UploadDuration   float64 `json:"upload_duration"`
-}
-
-type ImageRecognitionRequest struct {
-	Bucket      string `json:"bucket"`
-	Source      string `json:"source"`
-	ForceRemote bool   `json:"force_remote"`
-	ForceBackup bool   `json:"force_backup"`
-}
-
-type ImageRecognitionResponse struct {
-	Predictions       []string `json:"predictions"`
-	ForceRemote       bool     `json:"force_remote"`
-	ShortResult       bool     `json:"short_result"`
-	CodeDuration      float64  `json:"code_duration"`
-	DownloadDuration  float64  `json:"download_duration"`
-	InferenceDuration float64  `json:"inference_duration"`
-}
-
-type ImageScaleResult struct {
-	Duration float64            `json:"duration"`
-	Response ImageScaleResponse `json:"response"`
-}
-
-type ImageRecognitionResult struct {
-	Duration float64                  `json:"duration"`
-	Response ImageRecognitionResponse `json:"response"`
-}
-
-// ---------------------------------------
-type VideoSplitRequest struct {
-	Bucket      string `json:"bucket"`
-	Source      string `json:"source"`
-	Destination string `json:"destination"`
-	ForceRemote bool   `json:"force_remote"`
-}
-type VideoSplitResponse struct {
-	ForceRemote      bool    `json:"force_remote"`
-	CodeDuration     float64 `json:"code_duration"`
-	DownloadDuration float64 `json:"download_duration"`
-	SplitDuration    float64 `json:"split_duration"`
-	UploadDuration   float64 `json:"upload_duration"`
-}
-type VideoSplitResult struct {
-	Duration float64            `json:"duration"`
-	Response VideoSplitResponse `json:"response"`
-}
-
-type VideoTranscodeRequest struct {
-	Bucket      string `json:"bucket"`
-	Source      string `json:"source"`
-	Destination string `json:"destination"`
-	ForceRemote bool   `json:"force_remote"`
-}
-type VideoTranscodeResponse struct {
-	ForceRemote       bool    `json:"force_remote"`
-	CodeDuration      float64 `json:"code_duration"`
-	DownloadDuration  float64 `json:"download_duration"`
-	TranscodeDuration float64 `json:"transcode_duration"`
-	UploadDuration    float64 `json:"upload_duration"`
-}
-type VideoTranscodeResult struct {
-	Duration float64                `json:"duration"`
-	Response VideoTranscodeResponse `json:"response"`
-}
-
-type VideoMergeRequest struct {
-	Bucket      string `json:"bucket"`
-	Source      string `json:"source"`
-	Destination string `json:"destination"`
-	ForceRemote bool   `json:"force_remote"`
-}
-type VideoMergeResponse struct {
-	ForceRemote      bool    `json:"force_remote"`
-	CodeDuration     float64 `json:"code_duration"`
-	DownloadDuration float64 `json:"download_duration"`
-	MergeDuration    float64 `json:"merge_duration"`
-	UploadDuration   float64 `json:"upload_duration"`
-}
-type VideoMergeResult struct {
-	Duration float64            `json:"duration"`
-	Response VideoMergeResponse `json:"response"`
-}
-
-// ---------------------------------------------
-type FunctionChainResult struct {
-	StartTs  int64                  `json:"start_ts"`
-	Duration float64                `json:"duration"`
-	IsResult ImageScaleResult       `json:"is_result"`
-	IrResult ImageRecognitionResult `json:"ir_result"`
-	VsResult VideoSplitResult       `json:"vs_result"`
-	VtResult VideoTranscodeResult   `json:"vt_result"`
-	VmResult VideoMergeResult       `json:"vm_result"`
-}
 
 func function_chain(index int, bucket string, source string, forceRemote bool, useMem bool, workflowType string) FunctionChainResult {
 	fmt.Println("function_chain", index, "start")
